pyproject.toml
[tool.poetry]
name = "gigacontroller"
version = "0.2.0"
description = "GigaChat controller with grpc interface"
authors = ["Michael S2pac <mixa21.11@mail.ru>"]
readme = "README.md"
packages = [{ include = "gigacontroller", from = "src" }]

[tool.poetry.dependencies]
python = "^3.11"
loguru = "^0.7.2"
grpcio-status = "^1.60.1"
googleapis-common-protos = "^1.62.0"
gigachat = "^0.1.15"
python-dotenv = "^1.0.1"
grpcio = "^1.60.1"
grpcio-tools = "^1.60.1"
tenacity = "^8.2.3"
grpcio-health-checking = "^1.62.0"

[tool.poetry.group.dev.dependencies]
pylint = "^3.0.2"
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
ruff = "^0.1.6"
grpcio-testing = "^1.62.0"

[build-system]
requires = ["poetry-core>=1.6"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.scripts]
giga-controller = "gigacontroller.__main__:main"

[tool.black]
color = true
line-length = 120
target-version = ["py311"]

[tool.ruff]
extend-select = ["A", "B", "C4", "C90", "I", "Q", "PERF", "PT", "RUF100", "UP", "W"]
line-length = 120
src = ["src"]
target-version = "py311"

[tool.pylint.'MESSAGES CONTROL']
max-line-length = 120
disable = """
    too-many-arguments,
    not-callable,
    missing-class-docstring,
    missing-module-docstring,
    too-few-public-methods
"""

[tool.pytest.ini_options]
pythonpath = [
    ".", "src",
]
asyncio_mode = "auto"
addopts = "-s --verbose -W ignore::DeprecationWarning"